+++ toot/commands.py
@@ -8,7 +8,22 @@ from toot.utils import assert_domain_exists
 
 
 def timeline(app, user, args):
-    items = api.timeline_home(app, user)
+    # Make sure tag, list and public are not used simultaneously
+    if len([arg for arg in [args.tag, args.list, args.public] if arg]) > 1:
+        raise ConsoleError("Only one of --public, --tag, or --list can be used at one time.")
+
+    if args.local and not (args.public or args.tag):
+        raise ConsoleError("The --local option is only valid alongside --public or --tag.")
+
+    if args.public:
+        items = api.timeline_public(app, user, local=args.local)
+    elif args.tag:
+        items = api.timeline_tag(app, user, args.tag, local=args.local)
+    elif args.list:
+        items = api.timeline_list(app, user, args.list)
+    else:
+        items = api.timeline_home(app, user)
+
     print_timeline(items)